+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# 2026-01-08
+- #2312 Implements `eth_feeHistory` RPC endpoint to expose rollup's EIP-1559 base fee history for wallets.
+
 # 2026-01-02
 - #2292 **EVM Breaking Change, Chain Hash Change**. This PR Changes the borsh serialization of sov_evm::CallMessage by making it an enum. After upgrading, chains will not be able to deserialize transactions in the old format - so it will become impossible to sync from genesis if your chain pre-dates this change.
 # 2026-01-05